+++ semcore/d3-chart/src/Bar.js
@@ -1,9 +1,10 @@
 import React from 'react';
+import { transition } from 'd3-transition';
 import { Component, sstyled } from '@semcore/core';
+import uniqueIDEnhancement from '@semcore/utils/lib/uniqueID';
 import createElement from './createElement';
 import ClipPath from './ClipPath';
-import uniqueIDEnhancement from '@semcore/utils/lib/uniqueID';
-import { transition } from 'd3-transition';
+import { getBandwidth, roundedPath } from './utils';
 
 import style from './style/bar.shadow.css';
 
@@ -16,6 +17,7 @@ class BarRoot extends Component {
     color: '#50aef4',
     offset: [0, 0],
     duration: 500,
+    r: 2,
   };
 
   getBackgroundProps(props, index) {
@@ -50,28 +52,54 @@ class BarRoot extends Component {
 
   renderBar(d, i) {
     const SBar = this.Element;
-    const { styles, color, x, y, y0, scale, hide, offset, duration, uid } = this.asProps;
+    const {
+      styles,
+      color,
+      x,
+      y,
+      y0,
+      scale,
+      hide,
+      offset,
+      duration,
+      uid,
+      r,
+      $index,
+      width: widthProps,
+    } = this.asProps;
 
     const [xScale, yScale] = scale;
+    const barY = yScale(Math.max(d[y0] ?? 0, d[y])) + offset[1];
+    const barX = xScale(d[x]) + offset[0];
+    const height = Math.abs(
+      yScale(d[y]) - Math.min(yScale(yScale.domain()[0]), yScale(d[y0] ?? 0)),
+    );
+    const width = widthProps || getBandwidth(xScale);
+    const isRounded = r !== 0;
 
     return sstyled(styles)(
-      <SBar
-        key={`bar-${i}`}
-        render="rect"
-        clipPath={`url(#${uid})`}
-        __excludeProps={['data', 'scale', 'value']}
-        childrenPosition="above"
-        value={d}
-        index={i}
-        hide={hide}
-        color={color}
-        // TODO: https://github.com/airbnb/visx/blob/2fa674e7d7fdc9cffea13e8bf644d46dd6f0db5b/packages/visx-shape/src/util/getBandwidth.ts#L3
-        width={xScale.bandwidth()}
-        height={Math.abs(yScale(d[y]) - Math.min(yScale(yScale.domain()[0]), yScale(d[y0] ?? 0)))}
-        x={xScale(d[x]) + offset[0]}
-        y={yScale(Math.max(d[y0] ?? 0, d[y])) + offset[1]}
-        use:duration={`${duration}ms`}
-      />,
+      <>
+        <SBar
+          key={`bar-${i}`}
+          render="path"
+          clipPath={`url(#${uid})`}
+          __excludeProps={['data', 'scale', 'value']}
+          childrenPosition="above"
+          value={d}
+          index={i}
+          hide={hide}
+          color={color}
+          d={getRect({
+            x: barX,
+            y: isRounded ? (d[y] > 0 ? barY - r : barY) : barY,
+            width,
+            height: isRounded ? height + r : height,
+            radius: r,
+            position: d[y] > 0 ? 'top' : 'bottom',
+          })}
+          use:duration={`${duration}ms`}
+        />
+      </>,
     );
   }
   render() {
@@ -110,6 +138,16 @@ function Background(props) {
       y={yRange[1]}
     />,
   );
+}
+
+function getRect({ x, y, width, height, radius, position }) {
+  if (height <= radius) return '';
+  if (radius) {
+    if (position === 'top')
+      return roundedPath(x, y, width, height, radius, true, true, false, false);
+    return roundedPath(x, y, width, height, radius, false, false, true, true);
+  }
+  return roundedPath(x, y, width, height, radius);
 }
 
 export default createElement(BarRoot, { Background });