+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0] - 2021-09-30
+
+### Added
+
+- Added radius of curvature for `Bar, HorizontalBar`.
+- Added dynamic position tooltip for `HoverRect, HoverLine`.
+- Added hover for `Donut.Pie`.
+
 ## [1.2.0] - 2021-09-24
 
 ### Added